JSON Decode check, removed sleep timer
- catch JSON decoding errors when starting a new download
- don't sleep in the early morning anymore

The removal of the sleep timer may cause issues with long running scripts,
as the overnight sleep period was used to reset a counter that will now
continue to grow. There is no "max integer" in Python, but it still might
get unwieldy over time, since it does take the modulus quite often.

In other words, quit or kill the script overnight once in a while.

I'll add some proper overnight maintenance later.