Avoid dying on Showroom maintenance
These fixes have not been fully tested, but it *should* just output
two warnings at the start of maintence (maybe several if some rooms
are still broadcasting up to the start of maintenance.) and then
go quiet, and resume once its over.

It *should* sleep for a bit when it detects a maintenance-related error
(i.e. 503 errors), but the way I set it up, I can't change the interval.